Move task listener to own file as monitoring
